requirements.txt
langchain
langchain-community
langchain-openai
streamlit
faiss-cpu
python-dotenv
openai
tiktoken
sentence-transformers